move the config out of the extension so state isn't shared between generations
